Apply update of higher-order function function-type arguments contexts when they appear as optional
Summary:
We currently update the context of function-type arguments to higher-order functions (i.e. rewrite `Hfun_context '$X'`, where '$X' is the name of the function parameter, to a wildcard) when they appear at the top level but it not when the argument is optional. This diff pushes the existing update inside `Hoption` hints where `Hfun` may also appear with wildcard contexts.

Based on my understanding of where wildcard contexts may/must appear, we explicitly do not push the update inside other hints because one of the following applies:
- wildcard contexts are disallowed (soft types, like type , tuples, shapes)
- a function type cannot appear there (access, apply, prim)
- the type is not denotable

Reviewed By: madgen

Differential Revision: D26515236

fbshipit-source-id: b0fcab9c788808c59f52916b7e3a5b7fbe81e16f