Do not pass in string representation of symbol
Summary:
Used to pass a string representation of the function pointer. However, the name was not fully qualified, which sometimes causes name collision issues.

If frameworks want to get the name out, they may use reflection or one of the special function pointer get name methods to get a fully qualified/unique string that identifies the function.

Reviewed By: Wilfred

Differential Revision: D26597632

fbshipit-source-id: 8ebac49ba25e0fad55e1ab786337869abb66b4df