Tests to demonstrate use of MemoryManager::MaskAlloc by extensions.
Summary:
Demonstrate how extensions can use MemoryManager::MaskAlloc to 1) allocate memory on a request thread and free it in a background thread later or 2) allocate memory on a background thread and free it on a request thread later, in both cases without skewing the memory accounting on the request thread in question.

NOTE: using this for case 1 requires extension authors to add their own mechanism for ensuring the server's memory cannot grow without bound (e.g. limiting the size of the queue that tasks owning memory allocated in this way on request threads can sit in, and failing to enqueue more work if this server-wide queue limit is reached).

Reviewed By: binliu19

Differential Revision: D26593726

fbshipit-source-id: 6b043bd9a095c061dacdfa39a31ca0a164fa7334