Fix bug with reified type parameter arity verification
Summary:
When a function has reified generics, we check that arity of generics match, even for erased ones since we send `_` for their generics to the runtime.

The lowerer was converting context constants into erased generics for the typechecker. This convertion is incorrect for the runtime.

This diff removes such conversion from the lowerer when the code is emitting for runtime.

Reviewed By: vassilmladenov

Differential Revision: D26703935

fbshipit-source-id: 99ede475c3b1de8ef750eaa07a9dda09fc0a1d89