Rename DataWalker's hasObjectOrResource
Summary: We're using it to test for things that can't be made persistent (non-persistent arrays are accepted because we can make persistent copies), which rejects a lot more than objects and resources these days. Rename it to hasNonPersistable (hair-splitting: -able and not -ent because we accept things we can make persistent ourselves) to reflect what it's for.

Reviewed By: paulbiss

Differential Revision: D26679149

fbshipit-source-id: c4d7e7b707538d60f79eda90d00974caf577a439