Helpers to convert from Pos.t to Pos_or_decl.t
Summary:
These helpers will be used in the next diff.
We define a type Decl_reference which we aim at replacing the filename in positions, and should contain just enough information to be able to retrieve the file names from the naming tables.
The helpers in Decl_env use the existing env.Decl_env.root field, which is a Typing_deps.Deps.dep which we can easily convert to a Decl_reference.t.

Reviewed By: ljw1004

Differential Revision: D26339025

fbshipit-source-id: c27c5fdc9daf3f96f372994d419b4b0560c619d9