Fix probability leak in StaticStrMonotypeDict
Summary:
If we do an Elem/Set/Remove op on a StaticStrDict, even if we use a StringDict native helper, we should keep the layout index as tight as possible.

A bug here could cause performance issues, if we specialize a setter on StringDict but a downstream getter on StaticStrDict, but it doesn't seem to cause trouble in practice.

Reviewed By: colavitam

Differential Revision: D26744444

fbshipit-source-id: 990d7d0a2fa949017ed7b883b377fb3e10177826