Secondary PHP backtrace collection for crash handler
Summary:
Currently, if we segfault (or raise another unexpected signal) from the TC, we are unable to collect a PHP stacktrace if a fixup is missing. This leads to a significant number of cores lacking a PHP stacktrace which are instead attributed to "null".

In many cases, we can construct a PHP backtrace by scanning the stack for the first VM frame, faking the vmpc, and requesting a backtrace starting from this frame. Add this functionality as a secondary implementation of the backtrace routine in the crash handler.

If the secondary routine fails, we simply signal again and skip it.

Reviewed By: oulgen, jano

Differential Revision: D26743274

fbshipit-source-id: 955ef12963eb2a8946ef9d94d89cf4788950ec6e