Reduce cost of bespoke vdispatch
Summary: Modify bespoke dispatch so it's done off of a global table indexed by the high byte of the bespoke array's layout index, rather than off of a per-layout vtable. Doing so saves a pointer chase and a few arithmetic instructions for each operation. We must also add a new EmptyMonotypeVec layout byte, but luckily, we have space to do so.

Reviewed By: colavitam

Differential Revision: D26755011

fbshipit-source-id: 85506487478de6be4d50b427e12be2ce4d835cfb