Make ty_expr exhaustive on expressions
Summary:
This diff makes ty_expr exhaustive, so that we know we're covering every case of readonlyness on expressions.
It also adds a large number of tests for the cases we were missing before; none of them are really that interesting so I didn't split the diff up, but let me know if there's any you'd like me to split out.

Reviewed By: zilberstein

Differential Revision: D26593993

fbshipit-source-id: 2b5ba67cad0e52116b7f6439ef3471e0f9ad50a1