Remove lazy_emit_state
Summary: We are working towards integrating a new version of `InstrSeq` that makes use of arena allocation strategy. Unfortunately `lazy_emit_state` can't be implemented in that regime. It seems it can't be done in safe Rust (for `$type` types `T<'arena>`), at least not for lifetimes other than `'static`. This change removes the macro and substitutes handwritten code to replace the functionality it used to generate.

Reviewed By: shiqicao

Differential Revision: D26795893

fbshipit-source-id: 827a4b59eb5b924159e777d1e3860fef38db3494