Migrate emitters to new `InstrSeq`
Summary:
We are working towards new `emit_*` functions that take advantage of the recently landed `InstrSeq<'arena>` types.

The step taken in this diff is to prepare the way by adapting all the `ast*.rs` files and all but two of the `hhas_*.rs` files (the remaining two i.e. `hhas_class.rs` and `hhas_constant.rs` are in a recursive cycle with `emit_*.rs` files and will get adapted in a later phase).

Reviewed By: shiqicao

Differential Revision: D26728155

fbshipit-source-id: c1f476507a31cb4ca4a85ebab1a8186c9cf78bb8