Fix infinite recursion in HDF parser
Summary:
It's possible to create a Node as a "link" (whatever that's
supposed to be) and then reference itself while still inside the
Node. This causes the lookup to go into an infinite loop.

I can only speculate as to the desired behavior here (this format has
no surviving documentation or commenting). Do the brute force thing
and add a recursion check to _walk_hdf. If we exceed 1000, return an
error. This was a bit awkward because _walk_hdf previously couldn't
return an error like this, so I had to thread it through a number of
places.

Add a gtest which triggers the recursion.

Reviewed By: jano

Differential Revision: D26799491

fbshipit-source-id: edf3c7791cf1186293f6e2b41605e6e4ef2f5ee6