adding reference to constant initializers in decl
Summary:
In order to correctly fix the unsounded caused by cyclic constant definitions,
we need to add constant initializers expressions back into the decl. In order
to keep things tracktable, we don't need the full expression but only
references to other constants that appear in the initializer expression.

This diff only perform this addition, do not use it yet. I'd like to see
how this impact the decl size first

Differential Revision: D26577250

fbshipit-source-id: 0bf2e16deeba0acd6411caae92a0df1e264c34b4