Reflect new switch semantics in the typechecker
Summary:
When no case matches the value, HHVM throws a runtime exception.
Take this new behavior into account during typing.

Note: this is subtly different from saying that there is an implicit
`default: throw new RuntimeException();` at the end, because if the
last 'case' falls through, it'll fall through *outside* the switch
and not in the implicit final default.

The diff happens to fix the following type hole:
```
function f(): int {
	$x = 0;
	switch(1) {
	case 0:
		$x = 10;
		break;
	case 1:
		$x = 'hi';
	}
	return $x;
}
```

Reviewed By: DavidSnider, CatherineGasnier

Differential Revision: D26842782

fbshipit-source-id: 2cda55b631aa96cab2466477c05f704e48612d22