Add a thread local to check whether we are jitting
Summary: In FSM we use whether g_context is null or not to detect whether it is safe to raise the notice but this is not enough. If we are doing a live translation, then g_context will not be null since we are using the request thread to jit.

Differential Revision: D26850528

fbshipit-source-id: f8eb2ef74091c35390a5c6b2ea6acdf767511637