Simplify calculation of the number of stack elements in EndCatch load elim
Summary:
We already have a FPInvOffset inside the EndCatch's marker, which is the offset
from the innermost stack base. The stack base of non-resumed functions
currently contains the frame slots, deduct that to obtain the number of stack
elements.

Replaces a bizarre computation that first abused AStack implementation to
compute the number of stack cells from the outermost frame, but skipping a hole
representing the innermost locals, and then trying to remove everything above
BeginInlining, but using a wrong frame of reference so adjusting for that while
adjusting for inlining.

Reviewed By: oulgen

Differential Revision: D25746259

fbshipit-source-id: 923ca432e6bd3f86201620d0be1ea818423c67a7