Rip out now-dead DVOverrides logic
Summary:
We tracked these DVOverrides in order to maintain the text "varray[]" for a default param value after we flipped the HADVAs switch. Some of our Hack codegen logic is quite sensitive to default values, to the point where user-land unit tests will fail if we change this format (oops...) The way DVOverrides are supposed to work is that hackc maintains the varray/vec distinction even post-HADVAs. We deserialize varrays to vecs in as.cpp, but then, in order to reconstruct the text, we track places where we should re-serialize vecs to the string "varray". It's all very fragile.

We foolishly did not write unit tests for this default-value behavior, so at some point, a hackc change made it so that we no longer maintain the distinction. The DVOverrides logic was still present in HHVM, but all the override bits were false. Recently, in D25930605 (https://github.com/facebook/hhvm/commit/a0e34393fc3e5e823a8d2b06ddd26475397fbc69), we broke indexing into this overrides vector (the key Offset used to be unique across a Unit but now it's only unique across a Func). Since it's dead, let's just delete it.

Reviewed By: jano, dneiter

Differential Revision: D26870073

fbshipit-source-id: 431ce325fb6c169dc1a340cd66028cc5dce5d7d2