Merge EagerSyncVMRegs into EnterTCUnwind
Summary:
As implemented, EnterTCUnwind requires that the VM registers are synced, so
EagerSyncVMRegs should be treated as an inseparable part of EnterTCUnwind.

For example, it would be illegal to reorder operations and insert code that
e.g. reenters VM, or pushes InlineCall in between these two operations, but the
memory effects of EagerSyncVMRegs allows such reordering to happen.

Fix this potential issue by merging EagerSyncVMRegs into EnterTCUnwind. Also
use inst.src(1) instead of inst.marker().fp() when referring to locals, which
is safe wrt inlining (in case we optimized away the frame, which we can't yet).

The main motivation of this diff was to make stack effects SP rather than FP
relative, as FP-relative addressing of stack elements is being removed.

Reviewed By: oulgen

Differential Revision: D26871452

fbshipit-source-id: 3cc0b3ae8052dee0e159a08490fda264d6934c9d