deps_mode 3: get deps_mode from cursor, not global
Summary:
My goal is to enable 64bit for hh_fanout. This means passing the correct deps_mode where it's needed. But we can only construct 64bit deps_mode after we've loaded saved-state. This is a sequence of diffs to allow deps_mode information to flow from the saved-state-load through to the places that need them.

Earlier diffs already made deps_mode flow from saved-state-load into fanout-state-cursor. This diff changes a few places that used to use global SQLite deps_mode but which have access to the cursor, to read it from them.

Differential Revision: D26770966

fbshipit-source-id: caf254a56cdf76348af4c280a02ba0cbd4798dca