bypass server_env
Summary:
This is a small self-contained refactoring. I noticed that we're constructing
```
server_env =
    ServerEnvBuild.make_env ~init_id ~deps_mode genv.ServerEnv.config
```
but then we're discarding everything in server_env other than the tcopt and popt that we already passed into it, in genv.config.

So I'm deleting server_env.

(I have an ulterior motive... I want to eliminate the global "deps_mode" variable, to allow me to move over to 64bit, and this is one place that uses it).

Differential Revision: D26770232

fbshipit-source-id: 316059da8031d363aa2a764878feba3d57835511