allow 64bit
Summary:
This diff adds a temporary flag --64bit to some hh_fanout commands, telling them to use the 64bit depgraph.

I want to remove 32bit-depgraph support from hh_fanout. That's because hh_fanout's 32bit depgraph is the ONLY place in our codebase which constraints the hash format inside the naming table - it forces it to have lower-32-bits for depgraph hash, and upper-32-bits for further name hash. I want to change the naming table to only use the same hashes as 64bit-depgraph.

I can't change the naming-table while hh_fanout clings on to its 32bit depgraph.

Therefore, I intend to force hh_fanout to use 64bit depgraph, with no option to go back.

That won't happen yet in this diff. This diff merely adds an option for now, so I can do manual tests side-by-side of 32bit and 64bit. I'll force the switch in the next diff in the stack.

Differential Revision: D26771033

fbshipit-source-id: d80e2ec513f60b4e0634323b897e7c8f803a4fec