deps_mode 1: return deps_mode from load_saved_state
Summary:
My goal is to enable 64bit for hh_fanout. This means passing the correct deps_mode where it's needed. But we can only construct 64bit deps_mode after we've loaded saved-state. This is a sequence of diffs to allow deps_mode information to flow from the saved-state-load through to the places that need them.

This diff adds deps_mode to the save_state_results structure. And, it threads that into two places that need deps_mode. (there are still other places that need deps_mode, which I'll fix up in future diffs).

Differential Revision: D26770628

fbshipit-source-id: bf1ebba168903968eca7361b5733696d64fbdc0f