deps_mode_ctx 1: rename previous_cursor -> cursor_reference
Summary:
This diff merely renames. Formerly:
```
type previous_cursor =
  | Previous_cursor_from_saved_state of saved_state_result
  | Previous_cursor_id of string
```
That thing wasn't a cursor! Cursors have the type Incremental.cursor and are a lot richer. That thing is solely a *reference* to a cursor, a reference that has to be resolved.

Why am I doing this? My overall goal is to make deps_mode and ctx flow correctly: (1) first you load the deptable, (2) next you construct ctx based on the deptable path and store deps_mode inside ctx, (3) next you load the naming-table, (4) the deps_mode result of loading the saved-state gets stored inside a cursor, (5) anyone who needs a ctx must get it either from loading a saved-state, or from a cursor.

This diff is related to step (5): if someone starts from a cursor, they're going to have to resolve/load that cursor first, get the deps_mode from it, and construct a ctx out of that.

Differential Revision: D26819307

fbshipit-source-id: 7d87f832c1b62c5b59fb222d0dc77e0b119f0a37