Implement StructArray and StructLayout for bespoke tests
Summary:
A `StructLayout` is a map from static string keys to slots. We have a single layout for a given set of keys.

A `StructArray` contains an array of `TypedValues`. The corresponding layout represents a superset of keys the array might have. If a key is not currently present in the array, the corresponding slot contains `UnInit`.
Additionally, the `StructArray` contains 64 bits of data to keep track of the insertion order. Currently, we can have <= 16 keys in a `StructArray` and the insertion order data contains 4 bits for each slot, arranged from first to last insertion position, the first one being at bits `m_size *4` to `(m_size - 1) * 4`.

Differential Revision: D26854394

fbshipit-source-id: 754e9a0cfb920bb351f5ebd21d7f10472a64476e