Add type info to XHP attributes (take 2)
Summary:
Note: Due to some rebase mishaps and a new approach, it was easier to start fresh. See D26553576 for previous discussions.

This diff adds type information to the `Xhp_simple` node, to support showing the correct type in the hover tooltip for XHP attributes in constructors (i.e. `<foo:bar my-attribute={...} .../>`).

We already have the type information in `'hi`, so we simply create a new record to store it.

Reviewed By: Wilfred

Differential Revision: D26794404

fbshipit-source-id: 35adf53e5453ad902b058a2a5ce4f9ed1d32fe1e