remove extra/confusing error message when unknown constant is passed
Summary:
Consider the following code snippet:
```
enum class EE : int {
  int A = 12;
}

function f(<<__Atom>> HH\MemberOf<EE, int> $x) : int {
  return $x;
}

<<__EntryPoint>>
function main(): void {
  f(#A);
  f(#B); // !LINE
}
```
Prior to this diff, `f(#B)` would produce the following errors:
```
  [15:3-8] Unknown constant B in EE
  [15:5-7] Atoms are not allowed in this position. They are only allowed in function call, if the function parameter is annotated with __Atom
```

The second error, "Atoms are not allowed in this position...", seems misleading as the error is not that atoms cannot be used there (`f(#A)` is valid), just that `#B` doesn't exist in this context.

This diff removes the second error message in the unknown constant cases.

Reviewed By: Wilfred

Differential Revision: D26764838

fbshipit-source-id: bbb9807c3d5a6958ef0b180381519f5fcfda4aac