Use IRSPRelOffset instead of FPRelOffset in AStack's internals
Summary:
FPRelOffset was a weird choice for tracking stack positions, as there may be
multiple frames in the translation and they may not even live on the stack. The
meaning of FPRelOffset in this case was heavily overloaded -- it was an offset
relative to the stack base of the top frame.

Use IRSPRelOffset instead, which is already given by most of the users. The
only exceptions are places where the value already came from AStack's high/low
bounds, and a few places in frame-state.cpp, which were broken for inlined
code, as they incorrectly used FPRelOffsets relative to the current frame.

This diff fixes them.

Reviewed By: ricklavoie, oulgen

Differential Revision: D26878633

fbshipit-source-id: 68042a81e04bf60fdabdef7f550b48df387d4fde