Publish 'gradual modularity' document as a HIP
Summary:
This is not actively being worked on, however it often comes up when discussing the problems it's
meant to solve; this should be somewhere public for reference, and the HIPs directory seems like the best place.

This would need updating to the form of other HIPs if it were pursued.

Reviewed By: jjergus

Differential Revision: D25408089

fbshipit-source-id: 77e1b3bc321b0382a5ab2b2a8d07607ade5dfb6f