Ban local contexts on ctx constants
Summary:
Having local contexts on ctx constants doesn't make sense as they cannot be converted to proper runtime representation.

In the JIT, we have an instruction to fetch the value of the context constant. Note that the runtime does not have a way to represent _local contexts. We convert it towards either defaults or shallow based on whether we want to read it for ambient coeffects or required coeffects. Since we do not know the value of a coeffects constant statically at jit time (since we might know the LSB class), we always need to convert it to one direction.

Reviewed By: vassilmladenov

Differential Revision: D26891488

fbshipit-source-id: 13cb2382c732685100307f2b46bfe81fb31e87aa