Improve how coercion deals with enforcement
Summary:
We want things of dynamic type to be able to flow through points where a user-supplied type is enforced, such as function parameters and returns. Previously, a variant of sub-typing where dynamic is (somewhat confusingly) treated as bottom was used. This handled the situation where a variable of type int | dynamic was passed to a function that takes an int. int | dynamic <: int was checked, and by treating dynamic as bottom, the check passes. If the check fails, we still need to determine if a normal sub-type check would pass. For example, if we are checking int <: dynamic | string, then (only) treating dynamic as bottom would not be correct.

This diff changes things so that we use the usual sub-typing, with dynamic unioned onto the enforced type. For the above example, we would instead check int | dynamic <: int | dynamic, using the usual sub-typing. If it fails, based on the restricted on which types can be enforced, no further sub-type checks need to be carried out.

Also change the et_enforced field from a boolean to a variant type to record if a type is Enforced, Unenforced, or PartiallyEnforced. This is in preparation for handling types like vec<int> where the vec is enforced, even though the int isn't.

Reviewed By: vsiles

Differential Revision: D26820817

fbshipit-source-id: 2dfb284b8ee8505fe83c265f52aa269cd7225ff9