Allow classes to be passed to classname without coercion (part 1)
Summary:
In many cases, values passed to `classname` type-hints are used as classes. Until we have a type-hint for classes, we allow classes and lazy classes to be passed to `classname` type-hints without coercion. We added a runtime-option for raise notices when (lazy) classes are passed to `classname` type-hints.

We handle type-structures in part 2.

Reviewed By: paulbiss

Differential Revision: D25947690

fbshipit-source-id: bfd7ba6a9aa9a10ad2ba83bfa865fdc9ef0bf821