Forbid impurities in default args of callables except constants
Summary:
Enum class variants can unfortunately appear in undesired places such as default argument
initializers of a function/method/constructor.
This may well be in non-top-level positions, e.g.,
```
class NestedEnumClassAccess {
  public function __construct(
    private EnumClassWrapper $d = new EnumClassWrapper(MyClassEnum::A)
  )[] {}
}
```
warranting a *systematic* solution that isn't based on pattern-matching.
Similar problem cases arise when default argument is
a call to an impure function, method or constructor.

Therefore, when typing arguments bound to arguments (default being one of them),
use the coeffect that only allows access to statics.
This should be sound because no denotable context gives just the AccessStaticVariable capability,
therefore it is practically equivalent to having pure default argument initialization
(note that impure function calls cannot happen).

Also, delay the point at which we add a function's coeffects until the above is done,
in order to mitigate perf hit by this change due to shadowing of coeffects
in the local environment (lenv).  This, also required the check for memoization
to be moved further as it synthesizes a call to `IMemoizeParam`'s `getInstanceKey()`,
which, in turn, requires coeffects to be present in lenv.

Reviewed By: vsiles

Differential Revision: D26419508

fbshipit-source-id: 5bc05b9ec7351d128f8c98ca99d16794747c38bb