Check for private ancestor after target cache lookup
Summary:
Previously,
```
// see private_dispatch.php below
 P::go($c); // Surprise!
 P::go($d); // ok2
 P::go($c); // ok1
```
is what printed for opt builds with `-vEval.Jit=1`. We expect both calls to `P::go($c)` to be `Surprise!` following the rule that accessible private methods always win.

After we called `P::go($d)`, when we call `P::go($c)` we miss in the cache and choose a candidate function to be the function in the same method slot as the function in the cache. At this point, the method in the cache is `D::hello()`. The method in this slot for C is the public `C::hello()`. (This is how we build our method tables.)
However, C also has a private ancestor. This is the method we want! When we build our method tables, we also set the `hasPrivateAncestor` flag, so we can use this to guard against choosing the wrong method here.

The comment is wrong in that we *have not* "ruled out the case where dispatch should've gone to a private function with the same name". The previous check only checks if the previous function was private, and if the new class derives from the old class, we pick the same function.

Reviewed By: ricklavoie

Differential Revision: D26957988

fbshipit-source-id: de8718c4deb6019d58dc1e279b769c418a049bb5