Keep HADVAs off by default for tests
Summary:
Unit we've burned the HADVAs flag out of existence, we're going to want to test the full extent of array and type-system behavior - that is, we're going to want to have tests that use both PHP and Hack arrays.

Modify our runtime tests to force this flag off by default. This change has no effect today, but we need to land it to turn the flag-flip change green.

Reviewed By: ricklavoie

Differential Revision: D26980558

fbshipit-source-id: c0a350a275a96f4827e85c6cf6559b2e7485f554