Clean up HackArrDVArrMark
Summary:
The "automarking" testing strategy was quite successful - we were able to run post-HAM unit and integration tests several months before our Hack codebase was ready for the migration. It found a pile of bugs.

However, we can now run all unit and integration tests with HADVAs alone. It's better to do so, since that's our target configuration.

WARNING: This change affects post-HAM behavior in a small way. Specifically, after this change, in bytecode.cpp's unpack logic and in type-array.h's toDArray and toVArray casts, we no longer clear the mark post-HADVAs. Clearing the mark in these places is wrong. Instead, we should propagate the mark, like the pre-HADVAs code would do. Other places where we clear the mark, like on creating a new vec or dict or escalating a ClsMeth, don't suffer from this problem, because newly-allocated arrays are not marked.

Reviewed By: dneiter

Differential Revision: D26948465

fbshipit-source-id: 4e5ef37ad065c2cd9adc4eec633b2fbbd7804172