Add `hh_parse --generate-hhi`
Summary:

Starting to work on having HSL truly built-in; we want to keep it typechecked, but we don't want
the typechecker to always typecheck the implementation. If we generate HHIs, we can remove
the HH_FIXMEs around the PHP builtin calls, as we will no longer need the body to typecheck when using
the deregister_phpstdlib option.

ocamlformat and buck support todo before commit

Test plan:

```
dune runtest test/hhi/generate
```