Fix loose ends in unit tests
Summary:
Fix a couple of failures found by running all HHVM contbuilds in a post-HAM mode:
- Add some missing options files
- Fix some overly-specific .expectf.hadva regexes (matching the regular .expectf)
- Fix a real bug in the interaction between BespokeArray, ClsMeth, and HADVAs (yay...)

The reason for dropping the function name in the type error expects (e.g. array_merge) is because when we inline a builtin, we attribute the type error to the caller. This difference only shows up in the NO_PGO contbuild. We should fix that independently, but it happens both pre- and post-HAM.

Reviewed By: oulgen

Differential Revision: D26988433

fbshipit-source-id: 06cf6977bd31acc55e8da8f1506887fb220e4e63