Let users ignore files for type checking
Summary:
This diff adds a flag to hh_server that instructs it to use a user-provided filter when deciding which files need to be retypechecked.

When the user starts hh with `--config enable_type_check_filter_files=true`, before starting to type check a batch of files, it will try and read `~/.hack_type_check_files_filter` where it expects to find some include/exclude regex filters. Before starting the type check, hh will remove any file that doesn't pass the filters.

hh_server will still invalidate decls even for files that do not pass the filters. This ensures that the errors that are reported, are correct and consistent with the repo state.

hh_server will still type check

- files that are open in the IDE
- files that a user has edited or `touch`-ed

------

Reviewed By: vsiles

Differential Revision: D26913094

fbshipit-source-id: 6c5cbcb7431e8c49fdb336289333a26e2727f1d0