Support union and generic types in property accesses
Summary:
This allows us to support union, intersection and generic types for property accesses, which is enabled by the diff below.

For intersection and generic types, it suffices to find a single type in the intersection or upper bound for generics with a property, since all properties are invariant with respect to readonlyness (and, as far as I can tell, it's not possible to create an inhabited intersection type containing types that aren't related on the inheritance hierarchy). Thus, it shouldn't be possible for there to be a type which has both a mutable and readonly property of the same name.

Similarly for generic types, it suffices to find a single upper bound that has the property, as the readonlyness of the property must match between all upper bounds of the generic type for the constraints to be solveable.

Reviewed By: zilberstein

Differential Revision: D26660943

fbshipit-source-id: 0d09ee622efa157c498d1bfcacde1198ea5825ad