Refactor property elements to return a list
Summary:
This refactoring prepares to handle union and generic types for property accesses.

When you do a static property or regular property access on a union type, we'll need to return the list of all class elements it could be returning. This diff refactors the get_prop_elt code to get a list of elements instead of a single one.

Afterwards, in the diff above, we'll use the list to implement union type obj and class gets.

Reviewed By: zilberstein

Differential Revision: D26660944

fbshipit-source-id: b625abac0c225f5054f4c09213e8755a03da60a7